+++ src/DAO/FicheroObjetosMovimiento.java
@@ -19,16 +19,16 @@ import java.util.ArrayList;
  * @author jesus
  */
 public class FicheroObjetosMovimiento {
+
     private ArrayList<Movimiento> listaDeMovimientos;
     private Movimiento movimiento;
-    private boolean flagEmpleadoExistente;
 
     public FicheroObjetosMovimiento() {
         listaDeMovimientos = new ArrayList<Movimiento>();
     }
     private static FicheroObjetosMovimiento ficheroObjetosMovimiento = null;
 
-    public static FicheroObjetosMovimiento devolverFicherosEscritura() {
+    public static FicheroObjetosMovimiento devolver() {
 
         if (ficheroObjetosMovimiento != null) {
             return ficheroObjetosMovimiento;
@@ -37,45 +37,59 @@ public class FicheroObjetosMovimiento {
             return ficheroObjetosMovimiento;
         }
     }
-        public void escribirFicherosObjMovimiento( String numeroDeMovimiento,double Importe,String tipo,String fecha, String descripcion, String numeroDeEmpleado ) throws IOException, ClassNotFoundException
-        {
-        File f = new File("MovimientoObjeto.dat");
-        FileOutputStream fOS;
-        if (!f.exists()) {
-            f.createNewFile();
-        }
-        if (f.length() > 0) {
-            leerFicherosObjMovimiento();
-        }
-        fOS = new FileOutputStream(f, false);
-        ObjectOutputStream oOS = new ObjectOutputStream(fOS);
 
-        if (listaDeMovimientos == null) {
-            listaDeMovimientos = new ArrayList<Movimiento>();
+    public boolean escribirFicherosObjMovimiento(String numeroDeMovimiento, double Importe, String tipo, String fecha, String descripcion, String numeroDeEmpleado) throws IOException, ClassNotFoundException {
+        boolean existe = false;
+        ArrayList<Movimiento> listaComprobarMovimiento = new ArrayList();
+        listaComprobarMovimiento = leerFicherosObjMovimiento();
+        for (Movimiento m : listaComprobarMovimiento) {
+            if (m.getNumeroDeEmpleado().equals(numeroDeEmpleado)) {
+                if (m.getNumeroDeMovimiento().equals(numeroDeMovimiento)) {
+                    existe = true;
+                }
+            }
         }
-        
+        if (!existe) {
+            File f = new File("MovimientoObjeto.dat");
+            FileOutputStream fOS;
+            if (!f.exists()) {
+                f.createNewFile();
+            }
+            if (f.length() > 0) {
+                listaDeMovimientos.removeAll(listaDeMovimientos);
+                listaDeMovimientos=leerFicherosObjMovimiento();
+            }
+            fOS = new FileOutputStream(f, false);
+            ObjectOutputStream oOS = new ObjectOutputStream(fOS);
 
-        listaDeMovimientos.add(new Movimiento(numeroDeMovimiento, Importe, tipo, fecha, descripcion, numeroDeEmpleado));
-         
-        for (Movimiento c : listaDeMovimientos) {
-            oOS.writeObject(c);
+            if (listaDeMovimientos == null) {
+                listaDeMovimientos = new ArrayList<Movimiento>();
+            }
+
+            listaDeMovimientos.add(new Movimiento(numeroDeMovimiento, Importe, tipo, fecha, descripcion, numeroDeEmpleado));
+
+           
+                oOS.writeObject(listaDeMovimientos);
+            
+
+            oOS.close();
+            
         }
-
-        oOS.close();
+        return existe;
     }
-        
-         public ArrayList leerFicherosObjMovimiento() throws IOException, ClassNotFoundException {
 
-        File f = new File("./MovientoObjetos.dat");
+    public ArrayList leerFicherosObjMovimiento() throws IOException, ClassNotFoundException {
+
+        File f = new File("MovimientoObjeto.dat");
         if (!f.exists()) {
             f.createNewFile();
-        }
-
+        }listaDeMovimientos.removeAll(listaDeMovimientos);
+           if(f.length() > 0){
         ObjectInputStream ois = new ObjectInputStream(new FileInputStream(f));
 
-        listaDeMovimientos.removeAll(listaDeMovimientos);
+        
         listaDeMovimientos = (ArrayList< Movimiento>) ois.readObject();
-        ois.close();
+        ois.close();}
         return listaDeMovimientos;
     }
 }
